requirements.txt
antlr4-python3-runtime==4.8
certifi==2020.6.20
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.2
numpy==1.18.0
pandas==1.1.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0